[solver] Constrain object base addresses to their natural alignment
init_pointer_obj only constrained an object's base address when its type
carried an explicit "alignment" attribute, which just the clang frontend
sets (aggregates via padding.cpp, or an explicit alignas). Every other
object -- notably any scalar, and every object on the CBMC --binary path
-- had a free base address, so `(uintptr_t)&x % alignof(T) == 0` was
satisfiably false and produced a spurious counterexample (C11 6.2.8,
[basic.align]).

Constrain to the type's alignment instead, which reduces to the explicit
attribute when present and to natural alignment otherwise. Packed types
keep alignment 1 and so stay unconstrained, as before. This required
moving alignment() out of clang-c-frontend/padding.cpp into util, beside
type_byte_size(): it is a general type-layout query, and the solver must
not depend on a frontend.
