+++ regression/esbmc/natural_alignment/test.desc
@@ -1,0 +1,4 @@
+CORE
+main.c
+
+^VERIFICATION SUCCESSFUL$
