+++ regression/esbmc/natural_alignment_fail/test.desc
@@ -1,0 +1,4 @@
+CORE
+main.c
+
+^VERIFICATION FAILED$
