+++ src/clang-c-frontend/padding.cpp
@@ -11,93 +11,6 @@
 #include <util/std_expr.h>
 #include <util/type_byte_size.h>
 
-static std::size_t ext_int_representation_bytes(const typet &type)
-{
-  // We represent an ExtInt with the smallest integer type that can hold it
-  // TODO: This should be limited to a maximum of 8 bytes, but pointer analysis
-  // currently expects the fields to be aligned to the least power of 2 greater
-  // than the width in bytes
-  const std::size_t bits = string2integer(type.width().as_string()).to_uint64();
-
-  std::size_t result;
-  for (result = 1; bits > result * config.ansi_c.char_width; result *= 2)
-    ;
-
-  return result;
-}
-
-BigInt alignment(const typet &type, const namespacet &ns)
-{
-  // we need to consider a number of different cases:
-  // - alignment specified in the source, which will be recorded in
-  // "alignment"
-  // - alignment induced by packing ("The alignment of a member will
-  // be on a boundary that is either a multiple of n or a multiple of
-  // the size of the member, whichever is smaller."); both
-  // "alignment" and "packed" will be set
-  // - natural alignment, when neither "alignment" nor "packed"
-  // are set
-  // - dense packing with only "packed" set.
-
-  // is the alignment given?
-  const exprt &given_alignment =
-    static_cast<const exprt &>(type.find("alignment"));
-
-  BigInt a_int = 0;
-
-  // we trust it blindly, no matter how nonsensical
-  if (given_alignment.is_not_nil())
-    a_int = string2integer(given_alignment.cformat().as_string());
-
-  // alignment but no packing
-  if (a_int > 0 && !type.get_bool("packed"))
-    return a_int;
-  // no alignment, packing
-  else if (a_int == 0 && type.get_bool("packed"))
-    return 1;
-
-  // compute default
-  BigInt result = 0;
-
-  if (type.id() == typet::t_array)
-    result = alignment(type.subtype(), ns);
-  else if (type.id() == typet::t_struct || type.id() == typet::t_union)
-  {
-    result = 1;
-
-    // get the max
-    // (should really be the smallest common denominator)
-    for (const auto &c : to_struct_union_type(type).components())
-      result = std::max(result, alignment(c.type(), ns));
-  }
-  else if (type.get_bool("#bitfield"))
-  {
-    // we align these according to the 'underlying type'
-    result = alignment(type.subtype(), ns);
-  }
-  else if (type.get_bool("#extint"))
-    result = ext_int_representation_bytes(type);
-  else if (
-    type.id() == typet::t_unsignedbv || type.id() == typet::t_signedbv ||
-    type.id() == typet::t_fixedbv || type.id() == typet::t_floatbv ||
-    type.id() == typet::t_bool || type.id() == typet::t_pointer)
-  {
-    type2tc thetype = migrate_type(type);
-    result = type_byte_size(thetype);
-  }
-  else if (type.id() == typet::t_symbol)
-    result = alignment(ns.follow(type), ns);
-  else
-    result = 1;
-
-  // if an alignment had been provided and packing was requested, take
-  // the smallest alignment
-  if (a_int > 0 && a_int < result)
-    result = a_int;
-
-  return result;
-}
-
 static struct_typet::componentst::iterator pad_bit_field(
   struct_typet::componentst &components,
   struct_typet::componentst::iterator where,
