+++ src/clang-c-frontend/padding.h
@@ -8,8 +8,6 @@
 #include <util/namespace.h>
 #include <util/mp_arith.h>
 
-BigInt alignment(const typet &type, const namespacet &);
-
 void add_padding(typet &type, const namespacet &ns);
 
 #endif // CPROVER_ANSI_C_PADDING_H
